fix: address game and migration review feedback
