+++ src/app/(tabs)/index.tsx
@@ -1,3 +1,5 @@
 import { HomeScreen } from '../../screens/HomeScreen';
 
-export default HomeScreen;
+export default function TabsRoute() {
+  return <HomeScreen />;
+}
